# public/swagger/web/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Demo API",
        "version": "1.0"
    },
    "paths": {
...
        "/jobtitles/findAllByPaging": {
            "get": {
                "tags": [
                    "Job Titles"
                ],
                "summary": "Find all job titles by pagination",
                "operationId": "jobTitles_findAllByPaging",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search string",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_perPage",
                        "in": "query",
                        "description": "Per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_sort",
                        "in": "query",
                        "description": "Order items",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {}
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobTitlesFindAllByPaging"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
...
        "/jobtitles/update/{jobTitleId}": {
            "put": {
                "tags": [
                    "Job Titles"
                ],
                "summary": "Update job title",
                "operationId": "jobTitles_update",
                "parameters": [
                    {
                        "name": "jobTitleId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update customer",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JobTitleSave"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
...
        "/permissions/update/{permId}": {
            "put": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Update a permission",
                "operationId": "permissions_update",
                "requestBody": {
                    "description": "Update a permission",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PermissionSave"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
...
        "/users/delete/{userId}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete user",
                "operationId": "users_delete",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "user uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
        "/users/findAll": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Find all users for autocompleters",
                "operationId": "users_findAll",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonFindAll"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CommonFindAll": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "JobTitleSave": {
                "properties": {
                    "jobTitleId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "companyId": {
                        "$ref": "#/components/schemas/ObjectId"
                    },
                    "jobTitleName": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "JobTitlesFindAllByPaging": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/JobTitlesFindAllByPagingObject"
                        }
                    },
                    "page": {
                        "type": "integer"
                    },
                    "perPage": {
                        "type": "integer"
                    },
                    "totalPages": {
                        "type": "integer"
                    },
                    "totalItems": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "JobTitlesFindAllByPagingObject": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "companyId": {
                        "$ref": "#/components/schemas/ObjectId"
                    },
                    "jobTitleName": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ObjectId": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PermissionSave": {
                "properties": {
                    "permId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "moduleName": {
                        "type": "string"
                    },
                    "resource": {
                        "type": "string"
                    },
                    "action": {
                        "$ref": "#/components/schemas/ObjectId"
                    },
                    "route": {
                        "type": "string"
                    },
                    "method": {
                        "$ref": "#/components/schemas/ObjectId"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}